added custom-made time picker

+++ app/src/main/java/com/aranea/apps/zlatendab/util/AppUtil.java
@@ -12,18 +12,18 @@ import java.util.Locale;
  */
 public class AppUtil {
 
-  public static float convertDpToPixel(float dp, Context context){
+  public static int convertDpToPixel(float dp, Context context){
     Resources resources = context.getResources();
     DisplayMetrics metrics = resources.getDisplayMetrics();
     float px = dp * (metrics.densityDpi / 160f);
-    return px;
+    return (int) px;
   }
 
-  public static float convertPixelsToDp(float px, Context context){
+  public static int convertPixelsToDp(float px, Context context){
     Resources resources = context.getResources();
     DisplayMetrics metrics = resources.getDisplayMetrics();
     float dp = px / (metrics.densityDpi / 160f);
-    return dp;
+    return (int) dp;
   }
 
   public static void changeLocale(Context context, String languageCode) {
